+++ Tests/Integration/Caching/ArrayAccessCacheTest.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Pinq\Tests\Integration\Caching;
+
+use Pinq\Caching\ArrayAccessCacheAdapter;
+
+class ArrayAccessCacheTest extends CacheTest
+{
+    protected function setUpCache()
+    {
+        return new ArrayAccessCacheAdapter(new ArrayAccessCacheImplementation());
+    }
+}